requirements.txt
PyYAML
python-vlc
pafy
youtube_dl
pytube